% Generated by roxygen2 (4.0.0): do not edit by hand
\name{allRowsSortedCpp}
\alias{allRowsSortedCpp}
\title{Check that each row of a numeric matrix is sorted in strictly increasing order}
\usage{
allRowsSortedCpp(A)
}
\arguments{
\item{A}{A numeric matrix}
}
\value{
TRUE if each row is sorted in strictly increasing order, FALSE otherwise
}
\description{
Check that each row of a numeric matrix is sorted in strictly increasing order
}
\details{
allRowsSortedCpp is adapted from http://stackoverflow.com/a/7601857.
Strict inequalities are required in order to avoid m-tuples with repeated
positions (https://github.com/PeteHaitch/cometh/issues/8).
}
\keyword{internal}

