\name{assayNames}
\alias{assayNames}
\title{An internal function}
\usage{
assayNames(object)
}
\arguments{
  \item{object}{A CoMeth object}
}
\value{
A character vector wit the names of the assays
}
\description{
An internal function
}
\note{
Should be an internal function. Copied from bsseq package
}

