% Generated by roxygen2 (4.0.1): do not edit by hand
\name{betaCor}
\alias{betaCor}
\title{Compute correlations of pairs of beta-values.}
\usage{
betaCor(cometh, mls, ipd, nil = 0L, feature = NA, ignore_strand = FALSE,
  method = c("pearson", "kendall", "spearman"))
}
\arguments{
\item{cometh}{A CoMeth1 object.}

\item{mls}{A \code{\link{MethylationLociSet}} object
containing the positions of all methylation loci in the sample. This is
usually the set of all methylation loci in the reference genome.
If a \code{\link{MethylationLociSet}} is unavailable, then set to \code{NA},
in which case \code{betaCor} effectively assumes that \code{cometh} contains
all potential methylation loci (this is not recommended).}

\item{ipd}{An integer vector of IPDs. Correlations will only be computed
using pairs with an IPD found in \code{ipd}. The default value,
\code{NULL}, uses all observed IPDs. Setting \code{ipd = 0} will removes the
IPD strata (see examples).}

\item{nil}{The number of intervening loci allowed in each pair. The default,
\code{nil = 0}, means that only pairs made of adjacent loci are considered.
To use all pairs, set \code{nil = Inf} but \emph{this will take a very long
time (and probably crash) because there are so many combinations}.}

\item{feature}{A character. The name of the feature used to stratify the
correlations. The default, \code{feature = NA} means that no feature is
used. Otherwise, \code{feature} must be metadata of the m-tuples in the
\code{cometh} object, that is, one of \code{colnames(mcols(cometh))}.}

\item{ignore_strand}{logical. Should the strand of the methylation loci be
ignored (default: \code{FALSE})?}

\item{method}{A character string indicating which correlation coefficient
is to be computed. One of "pearson" (default), "kendall", or "spearman",
can be abbreviated. \emph{Currently, only "pearson" and "spearman" are
supported because "kendall" is very slow (see Note in
\code{\link[stats]{cor}}).}}
}
\value{
A \code{data.frame}. While the exact format depends on the
parameters used, the first column is the correlation, the second column is
the sample name and the remaining columns are \code{IPD} and
\code{features}, if relevant. Therefore, each row corresponds to the
correlation for a particular "sample-IPD-features" combination.
}
\description{
Given a \code{\link{CoMeth1}} object, \code{betaCor} computes the
correlations of pairs of beta-values. Correlations are computed per-sample
and are stratified by the \code{ipd} and \code{features} of the pairs of
methylation loci.
}
\note{
Correlations are computed using \code{\link[stats]{cor}},
with \code{use = "na.or.complete".}
}

