% Generated by roxygen2 (4.0.0): do not edit by hand
\name{candidateDuplicateMTuplesCpp}
\alias{candidateDuplicateMTuplesCpp}
\title{An internal helper function used in finding duplicate m-tuples.}
\usage{
candidateDuplicateMTuplesCpp(a, b, C)
}
\arguments{
\item{a}{An integer vector of length n. An integer representation of the
chromosome of each m-tuple.}

\item{b}{An integer vector of length n. An integer representation of the
strand of each m-tuple.}

\item{C}{An integer matrix with n rows. Each row represents the positions
of an m-tuple.}
}
\value{
A logical vector. TRUE if the corresponding row of \code{x} is a
potential duplicate and FALSE otherwise.
\emph{NOTE:} TRUE does not mean that it is a duplicate, merely that it is a
candidate. Further checking is required of these candidates, e.g. using the
\code{\link[base]{duplicated.array}} method.
}
\description{
The m-tuple should have already been converted to an integer representation,
namely an integer vector for the chromosome, an integer vector
for the strand and an integer matrix for the  positions.
}

