% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{cometh-package}
\alias{cometh-package}
\title{Analyse, manage and visualise co-methylation data.}
\description{
\pkg{cometh} provides tools for analysing, managing and visualising co-methylation
data. Loosely speaking, co-methylation is the correlation structure of DNA
methylation.
}
\details{
Please refer to the vignettes to see how to use the  \pkg{cometh} package.
}
\references{
TODO
}

