% Generated by roxygen2 (4.0.0): do not edit by hand
\name{compareMTuplesCpp}
\alias{compareMTuplesCpp}
\title{An internal function used to compare m-tuples.}
\usage{
compareMTuplesCpp(a, b, C)
}
\arguments{
\item{a}{An integer vector of length n. An integer representation of the difference in chromosomes of each m-tuple.}

\item{b}{An integer vector of length n. An integer representation of the difference in strand of each m-tuple.}

\item{C}{An integer matrix with n rows. Each row represents the difference in positions of each m-tuple.}
}
\value{
An integer vector where each element is the comparison of a pair
of m-tuples.
If the first m-tuple in the pair is "<" than the second m-tuple then the
return value for that element is < 0, if the first m-tuple in the pair is
"==" the second m-tuple then the return value is 0, and if the first
m-tuple is ">" that the second m-tuple then the return value is > 0.
}
\description{
An internal function used to compare m-tuples.
}
\details{
The m-tuple should have already been converted to an integer representation,
namely an integer vector for the difference in chromosome, an integer vector
for the difference in strand and an integer matrix for the difference in
positions.
}
\keyword{internal}

