% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fread.comethylation}
\alias{fread.comethylation}
\title{A faster version of \code{\link{read.comethylation}}}
\usage{
fread.comethylation(files, sample_names, methylation_types, seqinfo,
  verbose = FALSE)
}
\description{
Uses \code{data.table::fread} for faster reading of \code{comethylation}
\code{.tsv} output files. See https://github.com/PeteHaitch/cometh/issues/18
for benchmarking and a discussion of pros and cons of this approach.
}
\keyword{internal}

