\name{getCounts}
\alias{getCounts}
\title{Get the \code{counts} from a \code{\link{CoMeth}} object}
\usage{
getCounts(CoMeth)
}
\arguments{
  \item{CoMeth}{a \code{\link{CoMeth}} object}
}
\value{
A numeric matrix
}
\description{
Get the counts from a \code{\link{CoMeth}} object.
\code{counts} is a numeric matrix storing the number of
times each co-methylation pattern is observed at each
m-tuple.
}

