% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{methods}
\name{getCoverage,CoMeth-method}
\alias{getCoverage,CoMeth-method}
\title{Get the \code{coverage} from a \code{\link{CoMeth}} object.}
\usage{
\S4method{getCoverage}{CoMeth}(x)
}
\arguments{
\item{x}{A \code{\link{CoMeth}} object}
}
\value{
A numeric matrix. Each column of the matrix corresponds to a sample and each row to an m-tuple.
}
\description{
The \code{coverage} of a \code{\link{CoMeth}} object is the number of reads for each m-tuple, i.e. \code{rowSums(getCounts(CoMeth))}. The current implementation is slow.
}

