% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getPos}
\alias{getPos}
\title{Extract the genomic positions from an \code{MTuples} object.}
\usage{
getPos(x, ...)
}
\arguments{
\item{x}{An \code{\link{MTuples}} or \code{\link{CoMeth}} object.}
}
\value{
A numeric matrix. Each row of the matrix is an m-tuple and each
column is a position. Note that this function does not return the seqnames
of the m-tuples; use the \code{\link{seqnames}} getter to do this.
}
\description{
Extract the genomic positions from an \code{MTuples} object.
}

