% Generated by roxygen2 (4.0.1): do not edit by hand
\name{makeMLS}
\alias{makeMLS}
\title{Make a MethylationLociSet for a methylation type and a reference genome.}
\usage{
makeMLS(methylation_type, bsgenome)
}
\arguments{
\item{methylation_type}{A character. The possible values are "CG", "CHG",
"CHH" or "CNN" or multiple values specified as a character vector, e.g.
c("CG", "CHG") or c("CHG", "CG") for "CG/CHG" methylation.}

\item{bsgenome}{A \code{\link[BSgenome]{BSgenome}} object of the reference
genome.}

\item{...}{Other arguments passed to \code{\link{[Biostrings]matchPattern}}.}
}
\value{
A \link{MethylationLociSet} object.
}
\description{
Construct an object to store the location of all methylation loci of the
given type in a genome sequence. For example, find all CpGs in the human
reference genome (hg19). The returned object is a
\code{\link{MethylationLociSet}}.
}
\note{
This is basically a wrapper around the
\code{\link[Biostrings]{matchPattern}} function, where
\code{methylation_type} = \code{pattern} and \code{bsgenome} =
\code{subject}.
\emph{WARNING:} This will use a considerable amount of memory and take some
time for non-CG methylation.
}

