% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.mbias}
\alias{read.mbias}
\title{Read the M-bias.txt file produced from Bismark.}
\usage{
read.mbias(file)
}
\arguments{
\item{file}{The name of the M-bias file which the data are to be read from.}
}
\value{
A \code{\link{data.frame}}. The columns are:
\itemize{
 \item \code{read}: \code{SE} if the data are from single-end sequencing;
\code{R1} or \code{R2} if the data are read_1 or read_2 from paired-end
sequencing.
 \item \code{read_pos}: the position along the read.
 \item \code{methylation_type}: the methylation type. Either "CG", "CHG",
"CHH".
 \item \code{M}: the number of methylated base calls.
 \item \code{U}: the number of unmethylated base calls.
 \item \code{total}: the total number of base calls.
 \item \code{perc_M}: the percentage of methylated base calls.
}
}
\description{
Read the M-bias.txt file produced from Bismark and return as a
\code{\link[base]{data.frame}}.
}
\note{
\strong{WARNING}: The "position along the read" is only the same as
the "sequencing cycle" if the reads did not undergo any pre-alignment
trimming.
}
\examples{
cat("TODO")
}

