% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rowDiffsCpp}
\alias{rowDiffsCpp}
\title{Calculates differences for each row in a matrix.}
\usage{
rowDiffsCpp(A)
}
\arguments{
\item{A}{An integer \eqn{N \times K} matrix.}
}
\value{
An integer \eqn{N \times (K - 1)} matrix.
}
\description{
Calculates differences for each row in a matrix.
}
\keyword{internal}

