% Generated by roxygen2 (4.0.0): do not edit by hand
\name{rowSumsHashInternalCpp}
\alias{rowSumsHashInternalCpp}
\title{An internal helper function used in finding duplicate rows of an integer matrix.}
\usage{
rowSumsHashInternalCpp(x)
}
\arguments{
\item{x}{An integer matrix.}
}
\value{
A logical vector. TRUE if the corresponding row of \code{x} is a
potential duplicate and FALSE otherwise.
\emph{NOTE:} TRUE does not mean that it is a duplicate, merely that it could be.
Further checking is required of elements with TRUE, e.g. using the
\code{\link[base]{duplicated.array}} method.
}
\description{
An internal helper function used in finding duplicate rows of an integer matrix.
}
\details{
Not actually used in cometh package but written during development of mTuplesHashCpp.
}
\keyword{internal}

