% Generated by roxygen2 (4.0.1): do not edit by hand
\name{betaCor}
\alias{betaCor}
\title{Compute correlations of pairs of beta-values.}
\usage{
betaCor(cometh, mls, nil = 0L, ipd, feature, feature_name,
  same_feature = FALSE, ignore_strand = FALSE, method = c("pearson",
  "kendall", "spearman"))
}
\arguments{
\item{cometh}{A CoMeth1 object.}

\item{mls}{A \code{\link{MethylationLociSet}} object
containing the positions of all methylation loci in the sample. This is
usually the set of all methylation loci in the reference genome.
If a \code{\link{MethylationLociSet}} is unavailable, then set to \code{NA},
in which case \code{betaCor} effectively assumes that \code{cometh} contains
all potential methylation loci (this is not recommended).}

\item{ipd}{An integer vector of IPDs. Correlations will only be computed
using pairs with an IPD found in \code{ipd}. If \code{ipd} is specified, then
the number of intervening loci (NIL) is ignore. Only one of \code{ipd} or
\code{nil} should be specified.}

\item{nil}{The number of intervening loci (NIL) allowed in each pair. The
default, \code{nil = 0}, means that only pairs made of adjacent loci are
considered. See the \code{ipd} parameter for information on how to use all
pairs, regardless of NIL. Only one of \code{ipd} or \code{nil} should be
specified.}

\item{feature}{A \code{GRanges} object. If supplied, the correlations are
stratified by whether the methylation loci overlap a range in the supplied
\code{feature}. For example, correlations might be stratified by whether the
loci are from a CGI. The details of this stratification are controlled by the
\code{same_strand} parameter.}

\item{same_feature}{logical. If supplying a \code{feature}, should pairs of
methylation loci be required to be in the same feature (TRUE) or just in the
same feature class (FALSE). Two loci are in the same feature class if, for
example, they are both in CGIs but each loci is in a \emph{different} CGI.
This only affects pairs of loci that are in the feature, i.e., pairs in the
"gaps" betweens features need not be in the same "gap".}

\item{ignore_strand}{logical. Should the strand of the methylation loci be
ignored (default: \code{FALSE}).}

\item{method}{A character string indicating which correlation coefficient
is to be computed. One of "pearson" (default), "kendall", or "spearman",
can be abbreviated. \emph{Currently, only "pearson" and "spearman" are
supported because "kendall" is very slow (see Note in
\code{\link[stats]{cor}}).}}
}
\value{
A \code{data.frame}. While the exact format depends on the
parameters used, the first column is the correlation, the second column is
the sample name and the remaining columns are \code{IPD} and
\code{features}, if relevant. Therefore, each row corresponds to the
correlation for a particular "sample-IPD-features" combination.
}
\description{
Given a \code{\link{CoMeth1}} object, \code{betaCor} computes the
correlations of pairs of beta-values. Correlations are computed per-sample
and are stratified by the \code{strand}, \code{features} and \code{ipd} of
the pairs of methylation loci.
}
\note{
Correlations are computed using \code{\link[stats]{cor}},
with \code{use = "na.or.complete".}
}

