% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{class}
\name{CoMeth-class}
\alias{CoMeth}
\alias{CoMeth-class}
\alias{CoMeth1-class}
\alias{CoMeth2-class}
\alias{CoMeth3Plus-class}
\alias{compare,CoMeth,CoMeth-method}
\alias{getCoverage}
\alias{getIPD,CoMeth-method}
\alias{getPos,CoMeth-method}
\alias{order,CoMeth-method}
\alias{sampleNames,CoMeth-method}
\alias{sampleNames<-,CoMeth,ANY-method}
\alias{show,CoMeth-method}
\title{CoMeth class and constructor}
\usage{
CoMeth(assays = SimpleList(), rowData = MTuples(), colData = DataFrame(),
  exptData = SimpleList(), ..., verbose = FALSE)

getCoverage(x, ...)

\S4method{getPos}{CoMeth}(x)

\S4method{getIPD}{CoMeth}(x)

\S4method{sampleNames}{CoMeth}(object)

\S4method{sampleNames}{CoMeth,ANY}(object) <- value

\S4method{compare}{CoMeth,CoMeth}(x, y)

\S4method{order}{CoMeth}(..., na.last = TRUE, decreasing = FALSE)

\S4method{show}{CoMeth}(object)
}
\arguments{
\item{assays}{A \code{\link[IRanges]{SimpleList}} of matrix elements. All
elements of the list must have the same dimensions, and dimension names
(if present) must be consistent across elements and with the row names of
\code{rowData} and \code{colData}. See below for the required assays for
each subclass of \code{CoMeth}.}

\item{rowData}{A \code{\link{MTuples}} instance describing the m-tuples of
interest. Row names, if present, become the row names of the \code{CoMeth}.
The length of the \code{CoMeth} must equal the number of rows of the
matrices in assays.}

\item{colData}{A \code{\link[GenomicRanges]{DataFrame}} describing the samples.
Row names, which are required, become the column names of the \code{CoMeth}.
See below for additional details.}

\item{exptData}{An optional \code{\link[IRanges]{SimpleList}} of arbitrary
content describing the overall experiment.}

\item{...}{S4 methods \code{list} and \code{matrix}, arguments identical to
those of the \code{SimpleList} method.}

\item{verbose}{A \code{logical(1)} indicating whether messages about data
coercion during construction should be printed.}

\item{x}{A \code{\link{CoMeth}} object}

\item{object}{Object, possibly derived from class \code{eSet}.}

\item{value}{Character vector containing feature or sample names.}

\item{decreasing}{
    See \code{?base::\link[base]{sort}}.
  }

\item{na.last}{
    See \code{?base::\link[base]{order}} for a description of
    these arguments.
  }
}
\value{
The \code{CoMeth} constructor returns the appropriate subclass
depending on  the size of the m-tuples, i.e., the value of \eqn{m}.

A numeric matrix. Each column of the matrix corresponds to a sample
and each row to an m-tuple.
}
\description{
The \code{CoMeth} class is a virtual class extended from
\code{\link[GenomicRanges]{SummarizedExperiment}}. The subclasses,
\code{CoMeth1} (for 1-tuples), \code{CoMeth2} (for 2-tuples) and
\code{CoMeth3Plus} (for m-tuples, \eqn{m \ge 3}), are containers for
storing methylation patterns at methylation loci m-tuples.

Most users will use the \code{\link{read.comethylation}} function to
construct the appropriate \code{CoMeth} object from the output \code{.tsv}
file(s) produced by the \code{comethylation} Python software.

Get the sample-specfic sequencing coverage of each m-tuple in a
\code{\link{CoMeth}} object.

#

Compare \code{MTuples}.
}
\section{Differences to \code{\link[GenomicRanges]{SummarizedExperiment}}}{

There are several differences between
\code{\link[GenomicRanges]{SummarizedExperiment}} and \code{CoMeth}.

\subsection{\code{rowData}}{
A \code{CoMeth} object uses an \code{\link{MTuples}} object in the
\code{rowData} slot rather than a \code{\link{GRanges}} or
\code{GenomicRangesList} object.

Furthermore, the \code{MTuples} object must include
\code{\link[GenomeInfoDb]{Seqinfo}} of the reference genome. This
"reference genome" can accommodate a spiked-in unmethylated genome (normally
lambda phage) that is commonly used as a control in bisulfite-sequencing
experiments.}

\subsection{\code{assays}}{
The assays of a \code{CoMeth} object must include the counts of how many
times each methylation pattern is observed for that m-tuple in each sample.
For example, there are \eqn{2^1} possible methylation patterns at 1-tuples,
namely \code{M} and \code{U}; there are \eqn{2^2 = 4} possible methylation
patterns at 2-tuples, namely \code{MM}, \code{MU}, \code{UM} and \code{UU};
there are \eqn{2^3} possible patterns at 3-tuples, namely \code{MMM}, ...,
\code{'UUU'}. The \code{CoMeth} class enforces non-negative integer values
in these "counts" assays. The assays must be a \code{\link[IRanges]{SimpleList}}.

There are is an additional assay for each of the \code{CoMeth1} and
\code{CoMeth2} subclasses; the assays of a \code{CoMeth1} object also
include the \eqn{\beta-{\text{values}}}, \code{beta}, defined as
\eqn{\beta = \frac{M}{M + U}}; the assays of a \code{CoMeth2} object also
include the within-fragment comethylation, which is measured as a log-odds
ratio, \code{LOR}, and is defined as
\eqn{\log(\frac{(MM + 0.5)(UU + 0.5)}{(MU + 0.5)(UM + 0.5)})}. The
\code{CoMeth3Plus} contains no additional assays.}

\subsection{\code{colData}}{
Unlike a \code{\link[GenomicRanges]{SummarizedExperiment}}, a \code{CoMeth}
object requires the specification of \code{colData}. The row names of this
\code{colData}, which are required, must be the sample names and will be
used as the column names of the \code{CoMeth}. The \code{colData} must
include the methylation type as a column of the \code{colData} and use the
column name \code{methylation_type}. The \code{methylation_type} must be one
of \code{CG}, \code{CHG}, \code{CHH} or \code{CNN}, or some combination of
these separated by a '/', e.g., \code{CG/CHG}.}
}

\section{Coercion}{

\strong{TODO}: Insert help for any coercion methods.
}

\section{Accessors}{

\strong{TODO}: Insert help for any accessor methods.
}

\section{Splitting and combining}{

\strong{TODO}: Insert help for any splitting and combining methods.
}

\section{Subsetting}{

\strong{TODO}: Describe any subsetting methods.
}

\section{Filtering}{

\strong{TODO}: Describe any filtering methods.
}

\section{Methods based on findOverlaps}{

\strong{TODO} Insert help for any findOverlaps-based methods.
}

\section{Other methods}{

\strong{TODO}: Describe any other methods.
}

\section{For developers}{

Developers may wish to extend the \code{CoMeth} virtual class to other
specific m-tuples, e.g. \code{CoMeth7} to include an additional assay
that is specific to 7-tuples.
}
\examples{
cat("TODO")
}
\references{
See \url{http://www.github.com/PeteHaitch/comethylation} for the
\code{comethylation} Python software.
}
\seealso{
\code{\link{read.comethylation}} for a function to read in the
\code{comethylation} output \code{.tsv} file(s) and construct the
appropriate \code{CoMeth} object.

\code{\link[GenomicRanges]{SummarizedExperiment}} for the class that
\code{CoMeth} extends.
}

