% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.comethylation}
\alias{read.comethylation}
\title{Parse \code{tsv} output files from the Python software \code{comethylation}.}
\usage{
read.comethylation(files, sample_names, methylation_types, seqinfo,
  quiet = FALSE)
}
\arguments{
\item{files}{The \code{.tsv} files created by \code{comethylation}. These files may be compressed with gzip or bzip2.}

\item{sample_names}{The sample names of each file.}

\item{methylation_types}{A character vector with the type of methylation of each file.}

\item{quiet}{logical: if \code{FALSE} (default), \code{read.comethylation()} will print a line, saying how many items (m-tuples) have been read per file.}

\item{seqinfo}{A \code{\link[GenomicRanges]{Seqinfo}} object containing information about the reference genome of the sample.}
}
\value{
A \code{\link{CoMeth}} object
}
\description{
Read the \code{tsv} output files from the Python software \code{comethylation} and construct a \code{\link{CoMeth}} object.
}
\details{
All files should be of the same size m-tuples, e.g. all 2-tuples.
}
\note{
The compression of \code{file} is determined by the file extension: \code{.gz} for compressed with gzip and \code{.bz2} for compressed with bzip2. Otherwise the file is assumed to be uncompressed.
}
\examples{
cat("TODO")
}
\seealso{
\code{\link{CoMeth}}
}

