% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.comethylation}
\alias{read.comethylation}
\title{Read \code{tsv} output files from \code{comethylation} software.}
\usage{
read.comethylation(files, sample_names, methylation_types, seqinfo,
  verbose = FALSE)
}
\arguments{
\item{files}{The \code{.tsv} files created by \code{comethylation}. These
files may be compressed with gzip or bzip2.}

\item{sample_names}{The sample names of each file.}

\item{methylation_types}{A character vector with the type of methylation
type of each file.}

\item{verbose}{logical Default \code{FALSE}. If \code{TRUE},
\code{read.comethylation()} will report some output when reading the files
with \code{\link[data.table]{fread}} and may produce some output when
constructing the \code{CoMeth} object.}

\item{seqinfo}{A \code{\link[GenomicRanges]{Seqinfo}} object containing
information about the reference genome of the sample.}
}
\value{
A \code{\link{CoMeth}} object
}
\description{
Read the \code{tsv} output files from the Python software
\code{comethylation} and construct a single \code{\link{CoMeth}} object. All
files should be of the same size m-tuples, e.g. all 2-tuples. Files may have
different \code{methylation_type}, although this is not recommended. All
files should be mapped against the same reference genome, which is supplied
as the \code{seqinfo} argument.
}
\note{
The compression of \code{file} is determined by the file extension:
\code{.gz} for compressed with gzip and \code{.bz2} for compressed with
bzip2. Otherwise the file is assumed to be uncompressed.
}
\examples{
cat("TODO")
}
\seealso{
\code{\link{CoMeth}}
}

